+++ lib/src/commands/mp_add_line_segment_command.dart
@@ -38,7 +38,6 @@ class MPAddLineSegmentCommand extends MPCommand with MPPosCommandMixin {
       newLineSegment: newLineSegment,
       lineSegmentPositionInParent: lineSegmentPositionInParent,
     );
-    elementEditController.afterAddLineSegment(newLineSegment);
   }
 
   @override