+++ lib/src/state_machine/mp_th2_file_edit_state_machine/mp_th2_file_edit_state_moving_end_control_points.dart
@@ -134,10 +134,9 @@ class MPTH2FileEditStateMovingEndControlPoints extends MPTH2FileEditState
       toLineSegmentsMap: modifiedLineSegmentsMap,
       descriptionType: MPCommandDescriptionType.editLine,
     );
-
     th2FileEditController.execute(lineEditCommand);
-    selectionController.updateSelectedElementClone(selectedLine.mpID);
-    th2FileEditController.triggerEditLineRedraw();
+    elementEditController.updateControllersAfterLineSegmentChangesPerLine();
+    elementEditController.updateControllersAfterElementChanges();
     th2FileEditController.stateController.setState(
       MPTH2FileEditStateType.editSingleLine,
     );