+++ lib/src/state_machine/mp_th2_file_edit_state_machine/mp_th2_file_edit_state_select_non_empty_selection.dart
@@ -66,7 +66,7 @@ class MPTH2FileEditStateSelectNonEmptySelection extends MPTH2FileEditState
 
       if (clickedElementAlreadySelected) {
         if (shiftPressed) {
-          selectionController.removeSelectedElements(clickedElements);
+          selectionController.removeSelectedElementsByElements(clickedElements);
         }
 
         return Future.value();